Answering 1 d)
Hi, i'm not that familiar with python notebooks and stuff, but when it comes to solving 1d), the gradient is straight-forward it simply uses the chain rule.
For the hessian, remember it is (if we abuse the syntax a little bit) the gradient columnwise gradient of the tranpose of the gradient.
So take the gradient, tranpose it, and then take the gradient of each term. Get that and expand it along the column and you get your hessian. So column i of the hessian is gradient(g'(aTx)*a_i) i.e hessian is g''(aTx)aaT.

Please correct the mathjax if i didn't do it properly.
Thank you for your repo, helps me check my answers